refactor(Drawer): コード構造のクリーンアップを実装
- ヘルパー関数convertToPixelsをutils.tsに分離
- テーマ参照をテスト環境に対応した固定値に変更
- 不要なimportを削除
- ファイル構造を他のコンポーネントと統一
- 全38テストが正常に動作することを確認
