+++ src/components/Drawer/Drawer.tsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState, useRef, useCallback } from "react";
 import Portal from "../Portal";
 import ConfirmModal from "../ConfirmModal";
 import { DrawerProps, CloseReason } from "./types";
+import { convertToPixels } from "./utils";
 import {
   Backdrop,
   DrawerContainer,
@@ -11,40 +12,6 @@ import {
   StickyFooter,
 } from "./styled";
 
-// 相対値をpx値に変換するヘルパー関数
-const convertToPixels = (value: string | number, direction: "left" | "right" | "bottom"): number => {
-  if (typeof value === "number") {
-    return value; // 数値はpx扱い
-  }
-  
-  // 文字列の場合は単位チェック
-  const numericValue = parseFloat(value);
-  
-  if (value.includes("vw")) {
-    return (numericValue / 100) * window.innerWidth;
-  }
-  if (value.includes("vh")) {
-    return (numericValue / 100) * window.innerHeight;
-  }
-  if (value.includes("%")) {
-    // %の場合は親要素基準だが、ここではviewport基準として扱う
-    const isHorizontal = direction === "left" || direction === "right";
-    return (numericValue / 100) * (isHorizontal ? window.innerWidth : window.innerHeight);
-  }
-  if (value.includes("px")) {
-    return numericValue;
-  }
-  
-  // 単位なしの文字列は警告を出す
-  if (!isNaN(numericValue)) {
-    console.warn(`Drawer: 文字列サイズには単位を指定してください。"${value}" → "${value}px" を推奨`);
-    return numericValue; // 一応数値として扱う
-  }
-  
-  // 無効な値の場合はエラー警告とフォールバック
-  console.error(`Drawer: 無効なサイズ値です: "${value}". デフォルト値を使用します。`);
-  return 400; // フォールバック
-};
 
 const Drawer: React.FC<DrawerProps> = ({
   isOpen,
